feat(premium): add restoreSubscription to MockSubscriptionStore and move MockSubscriptionStore to its own file
